Refactor(string to array middleware): update middleware functionality
Description:
In order to make the string to array middleware function to be able to be used for data in other parts of the request such as params and queries, Here I have added a new parameter name `location` to function so the developer be able to define the location of the data that he/she needs to convert from string to an array.

Additional: I have added comment to function in order to increase its readability.

Change Logs:
[String to array middleware] [Modified] - Add a new parameter to function and change the functionality accordingly.
[String to array middleware] [Doc] - Add comments to middleware function.
[Admin blog router] [Modified] - Update the use case of the string to array middleware.
[Admin courses router] [Modified] - Update the use case of the string to array middleware.
[Admin products router] [Modified] - Update the use case of the string to array middleware.